backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
SQLAlchemy
psycopg2-binary
python-multipart
python-jose[cryptography]
bcrypt
python-dotenv
google-genai
